requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.12.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.13.0
distro==1.9.0
Flask==3.1.2
flatbuffers==25.12.19
frozendict==2.4.7
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mediapipe==0.10.31
multitasking==0.0.12
numpy==2.4.1
opencv-python==4.13.0.90
pandas==2.3.3
peewee==3.19.0
platformdirs==4.5.1
protobuf==6.33.4
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soupsieve==2.8.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
Werkzeug==3.1.5
yfinance==1.0
